pyproject.toml
[tool.poetry]
name = "mmdx"
version = "0.1.0"
description = ""
authors = ["Aécio Santos <aecio.solando@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.3.3"
lancedb = "^0.2.2"
numpy = "^1.25.2"
duckdb = "^0.8.1"
Pillow = "10.4.0"
pandas = "^2.0.3, !=2.1.0"
torch = ">=2.0.0, !=2.0.1"
pyarrow = "^13.0.0"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.0"
minio = "^7.1.17"
boto3 = "^1.28.83"
openai = "^1.60.2"
scikit-learn = "^1.6.1"
peft = "0.5.0"
dataset = "^1.6.2"
datasets = "2.11.0"
hdbscan = "^0.8.40"
ollama = "^0.4.7"
gensim = "^4.3.3"
nltk = "^3.9.1"
transformers = "4.39.0"
flask-socketio = "^5.5.1"

[tool.poetry.dev-dependencies]
pytest  = { version = "^6.2.4" }

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
